bugs and fixes:
cout asks for row everytime
-done -- changed the second cout.

selecting a seat goes from 0 to 10 instead of 0 to 9
-done -- just made the conditions from row > ROWS --> row > ROWS -1. same with col.

when user enters negative numbers for seat selection, it's getting the positive counterpart
 ex. if user enters -1, -1. the program selects 1,1
 I guess we can leave it here?
-done -- nothing to change. putting +1 is the same as 1.

the cancel option (2), prints "cancelled" even if it wasn't reserved
-done -- cout from the main file needed to be deleted/transfered to the implementation file.

the reservation option (1), reserves a seat even if it's already reserved
done -- same principle with the cancel bug.

after exiting option (3), rerunning the program resets the seating position:
-tbd -- might wanna talk to the prof for elaboration of the project.

notes:
Do we need to print the seats everytime we reserve or cancel?